refactor: extract functionality to pull token from header into 'extractAccessTokenFromRequest' util
